percentage type fix

+++ Server/BLL/DTO/DepositDTO.cs
@@ -11,7 +11,7 @@ namespace MoneyManager.BLL.DTO
 
         public DateTime To { get; set; }
 
-        public short Percentage { get; set; }
+        public float Percentage { get; set; }
 
         public decimal InitialAmount { get; set; }
     }
